update(teaching): update capture_the_flag task to include pwning
why this change? with the ctf workshop done this year and the (perceived) growing interest in cybersecurity this challenge needed updating to be slightly more challenging. the ctf workshop essentially already covered the contents of the task, which is why updating it with something not covered in the workshop is important

with this update I try to get them to explore rev and binary exploitation in more detail than earlier, and intend to challenge them with more involved concepts like shellcodes